Refactor detailed view to its own class and UI component
