+++ src/main/resources/view/DetailedContactPanel.fxml
@@ -4,6 +4,5 @@
 <?import javafx.scene.layout.VBox?>
 
 <VBox xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1">
-  <ListView fx:id="detailedContactViewPanel" prefHeight="0" VBox.vgrow="ALWAYS" />
-  <ListView fx:id="personListView" VBox.vgrow="SOMETIMES" />
+    <ListView fx:id="detailedContactViewPanel" VBox.vgrow="ALWAYS" />
 </VBox>
